requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
python-dotenv==1.0.0
SQLAlchemy==2.0.16
psycopg2-binary==2.9.9
Werkzeug==2.2.3
boto3==1.26.152
